MEstimator (with shifted median)
Improve mestimator:
- Separate class to compute Maximum Absolute Deviation
- Ability to shift the median manually, which could be is a nice
trick to use the mestimator on the full pointcloud (not done)

- Recomputes the MEstimator on the nereast neighbor cloud.
WRONG IN THIS COMMIT: in this commit, the median is shifted to
the estimated position. This is NOT needed, and makes the mestimator
useless, or worse, wrong in this case.

Here is what should be done instead:
- Find the nearest neighbors. This gives a noisy pointcloud on the real
  data containing the closest points to our current estimation.
  Compute the MEstimator on that cloud (estimate mad, hubert weights),
  and use it to weight the importance of points.

- The median shifting scheme can instead be used to automatically
  discard points on the full reference cloud that are < threshold
  (ie 2*maximum of residual error for instance).